requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pdf2image==1.16.3
google-cloud-aiplatform==1.36.4
Pillow==10.1.0
pydantic==2.5.2
python-dotenv==1.0.0
aiohttp>=3.8.0
asyncio>=3.4.3
tqdm>=4.65.0
